Refactored timers, added Catch2 tests for them - rewrote auto timer logic to use std::chromo - timer resolution is now microsecond - delta time is now in seconds with type float

+++ src/engine/ecs/system.hpp
@@ -22,7 +22,7 @@ namespace chestnut
         ISystem( CEngine& engine );
         virtual ~ISystem() = default;
 
-        virtual void update( uint32_t deltaTime ) = 0;
+        virtual void update( float deltaTime ) = 0;
     };
     
 } // namespace chestnut
